examples/nitz-demo/py/requirements.txt
pillow
click
requests
